Cholesky

Function to decompose weighting matrix.

Cholesky(x, treat, group1, group2 = NA, method = "MB", dimension)

Arguments

x

covariates.

treat

treatment indicator vector.

group1

see Details.

group2

see Details. Default: NA.

method

a string that takes values in "MB", "MB2", Default: 'MB'.

dimension

the dimension of covariate.

Details

group1 and group0

To estimate \(E(Y (1))\) (average treatment effect for group 1), you need to set group1 = 1 and ignore group2. Similarly, To estimate \(E(Y (0))\) (average treatment effect for group 0), you need to set group1 = 0 and ignore group2.

To estimate average treatment effect on the control group \(E(Y (1) | T = 0)\), you need to set group1 = 1 and group2 = 0. Similarly, To estimate average treatment effect on the treated group \(E(Y (0) | T = 1)\), you need to set group1 = 0 and group2 = 1.

This function is feasible when there are more than two groups.