Hierarchically Regularized Entropy Balancing

Hierarchically Regularized Entropy Balancing is a approximate balancing method that expands the feature space by including higher-order terms of covariates while imposes ridge penalties with a hierarchical structure on the higher-order terms.

HRB(
  covariate,
  treat,
  group1,
  group2 = NULL,
  outcome,
  alpha.space = c(0.1, 0.05, 0.01, 0.005, 0.001),
  iterations = 1000,
  convergence = 10^{
     -10
 },
  cv.fold = 5,
  second.moment = TRUE,
  third.moment = TRUE,
  interact = TRUE
)

Arguments

covariate

covariates.

treat

treatment indicator.

group1

see Details.

group2

see Details, Default: NULL.

outcome

outcome.

alpha.space

alpha.space is

iterations

iteration time in optimization problem, Default: 1000.

convergence

Criteria for determining convergence of optimization problems. Default: 10^-10.

cv.fold

cv.fold is

second.moment

second.moment is

third.moment

third.moment is

interact

interact is

Value

a HRB object with the following attributes:

AT: the estimate of average treatment effect in group1 (i.e, \(E(Y(group1))\)).

weight: the estimated univariate balancing weight.

GMIM: Generalized Multivariate Imbalance Measure that defines in our paper.

MAE: Mean Absolute Error.

alpha: the tuning parameter we choose.

Examples

# estimating ATE
set.seed(0521)
data <- si.data()
X <- data$X
treat <- data$Tr
Y <- data$Y

result1 <- HRB(covariate = X, treat = treat, group1 = 1, outcome = Y)
result2 <- HRB(covariate = X, treat = treat, group1 = 0, outcome = Y)

# an estimate of ATE
result1$AT - result2$AT
#>           [,1]
#> [1,] -1.699096

# estimating ATC
result3 <- HRB(covariate = X, treat = treat, group1 = 1, group2 = 0, outcome = Y)

# an estimate of ATC
result3$AT - mean(data$Y[data$Tr == 0])
#>           [,1]
#> [1,] -1.875974