Mahalanobis balancing

Mahalanobis balancing is a multivariate perspective of approximate covariate balancing method to estimate average treatment effect.

Usage

MB(
  covariate,
  treat,
  group1,
  group2 = NULL,
  outcome,
  opti.method = c("proximalC", "proximal", "BFGS"),
  method = c("MB", "MB2", "kernelMB"),
  delta.space = c(0.1, 0.05, 0.01, 0.005, 0.001),
  iterations = 1000,
  convergence = 10^{
     -8
 },
  rate = 10
)

Arguments

covariate

Covariates, it should be a n by d matrix where n is sample size and d is dimension.

treat

The vector of treatment assignments, it should be a vector of length n.

group1

group1 is a number, see Details.

group2

group2 is a number, see Details, Default: NULL.

outcome

The outcome vector, it should be a vector of length n.

opti.method

The optimization method that will be applied to optimize loss function. It should takes values in "BFGS", "proximal", "proximalC". Default: "proximalC".

method

Method to calculate weighting matrix. It should takes values in "MB", "MB2", "kernelMB". See Details. Default: "MB".

delta.space

Tuning parameter in balancing. See Details. Default: c(1e-1, 0.05, 1e-2, 0.005, 1e-3).

iterations

The maximum number of iteration times in optimization problem, Default: 1000.

convergence

The absolute tolerance used to determine a stopping criteria for the optimization problems. Default: 10^-8.

rate

The parameter in Tikhonov-regularized Newton update. Default: 10.

Value

a MB object with the following attributes:

AT: the estimate of average treatment effect in group1 (i.e, \(E(Y(group1))\)).

weight: the estimated Mahalanobis balancing weight.

GMIM: Generalized Multivariate Imbalance Measure that defines in our paper.

delta: the tuning parameter we choose.

Details

group1 and group0

To estimate \(E(Y (1))\) (average treatment effect for group 1), you need to set group1 = 1 and ignore group2. Similarly, To estimate \(E(Y (0))\) (average treatment effect for group 0), you need to set group1 = 0 and ignore group2.

To estimate average treatment effect on the control group \(E(Y (1) | T = 0)\), you need to set group1 = 1 and group2 = 0. Similarly, To estimate average treatment effect on the treated group \(E(Y (0) | T = 1)\), you need to set group1 = 0 and group2 = 1.

This function is feasible when there are more than two groups.

method can be a valid string, including

"MB": We choose the weighting matrix \({W}_1=[diag(\hat{\Sigma})]^{-1}\) where \(\hat{\Sigma}\) denotes sample covariance matrix.

"MB2": We choose the weighting matrix \({W}_2={[\hat{\Sigma}]^{-1}}\) where \(\hat{\Sigma}\) denotes sample covariance matrix.

"kernelMB": Firstly, we modify our covariate to \(X_i* = (\Phi(X_1,X_i), ..., \Phi(X_n,X_i))\), then we apply method "MB" to produce Mahalanobis balancing weights.

delta.space grid of values for the tuning parameter, a vector of candidate values for the degree of multivariate approximate covariate balance. The default is c(1e-1, 0.05, 1e-2, 0.005, 1e-3).

Examples

# estimating ATE
set.seed(0521)
data <- si.data()
X <- data$X
treat <- data$Tr
Y <- data$Y

result1 <- MB(covariate = X, treat = treat, group1 = 1, outcome = Y)
result2 <- MB(covariate = X, treat = treat, group1 = 0, outcome = Y)

# an estimate of ATE
result1$AT - result2$AT
#>             [,1]
#> [1,] -0.09062922

# estimating ATC
result3 <- MB(covariate = X, treat = treat, group1 = 1, group2 = 0, outcome = Y)

# an estimate of ATC
result3$AT - mean(data$Y[data$Tr == 0])
#>            [,1]
#> [1,] -0.2469105